Rename config to request
